Create a tabbable generic link
a tags still need an href for chrome (and others)
to make them able to be found while tabbing

however react complains about security of javascript;
as the href.

my fix is to simply make a basic a which has a default onclick
of prevent default.

we will still call the original on click